refactor: reorganize billing into smaller components, misc improvements
