+++ src/component/admin/billing/Billing.tsx
@@ -1,5 +1,4 @@
 import AdminMenu from '../menu/AdminMenu';
-import { BillingGrid } from './BillingGrid/BillingGrid';
 import { PageContent } from 'component/common/PageContent/PageContent';
 import { useContext } from 'react';
 import { ADMIN } from 'component/providers/AccessProvider/permissions';
@@ -8,9 +7,13 @@ import AccessContext from 'contexts/AccessContext';
 import { AdminAlert } from 'component/common/AdminAlert/AdminAlert';
 import { useInstanceStatus } from 'hooks/api/getters/useInstanceStatus/useInstanceStatus';
 import { Alert } from '@mui/material';
+import { BillingDashboard } from './BillingDashboard/BillingDashboard';
+import { BillingHistory } from './BillingHistory/BillingHistory';
+import useInvoices from 'hooks/api/getters/useInvoices/useInvoices';
 
 export const Billing = () => {
-    const { isBilling } = useInstanceStatus();
+    const { instanceStatus, isBilling } = useInstanceStatus();
+    const { invoices } = useInvoices();
     const { hasAccess } = useContext(AccessContext);
 
     return (
@@ -22,7 +25,14 @@ export const Billing = () => {
                     show={
                         <ConditionallyRender
                             condition={hasAccess(ADMIN)}
-                            show={() => <BillingGrid />}
+                            show={() => (
+                                <>
+                                    <BillingDashboard
+                                        instanceStatus={instanceStatus!}
+                                    />
+                                    <BillingHistory data={invoices} />
+                                </>
+                            )}
                             elseShow={() => <AdminAlert />}
                         />
                     }
